- refactor preprocessing code - remove utils.py, add preprocessing.py - clean up transformers.ipynb
